change location to int instead of float

+++ app/src/main/java/com/example/dungeoncrawler/model/Player.java
@@ -34,7 +34,7 @@ public class Player extends Entity implements Observable {
         this.direction = direction;
         this.score = score;
         this.name = name;
-        this.location = new Location(0.0f, 0.0f);
+        this.location = new Location(0, 0);
         this.difficulty = difficulty;
         wallList = new ArrayList<>();
     }
@@ -141,7 +141,7 @@ public class Player extends Entity implements Observable {
         return location;
     }
 
-    public void setLocation(float newX, float newY) {
+    public void setLocation(int newX, int newY) {
         location.setxCord(newX);
         location.setyCord(newY);
     }
